requirements.txt
Flask>=2.2.5
Flask-SQLAlchemy>=3.0.0
psycopg2-binary>=2.9.0
PyJWT>=2.8.0
gunicorn>=20.1.0
Flask-Cors>=3.0.10
pytest>=7.0.0
